Extension trait to add loading states to the Bevy app
