+++ bevy_asset_loader/assets/dynamic_asset.assets
