+++ bevy_asset_loader/examples/stageless_progress.rs
@@ -1,7 +1,7 @@
 use bevy::app::AppExit;
 use bevy::asset::LoadState;
 use bevy::prelude::*;
-use bevy_asset_loader::{AssetCollection, AssetLoader};
+use bevy_asset_loader::prelude::*;
 use iyes_loopless::prelude::*;
 use iyes_progress::{ProgressCounter, ProgressPlugin};
 
@@ -12,14 +12,15 @@ use iyes_progress::{ProgressCounter, ProgressPlugin};
 /// completes after one second. At that point, `iyes_progress` will continue to the next state
 /// and the app will terminate.
 fn main() {
-    let mut app = App::new();
-    app.add_loopless_state(MyStates::AssetLoading);
-    AssetLoader::new(MyStates::AssetLoading)
-        .continue_to_state(MyStates::Next)
-        .with_collection::<TextureAssets>()
-        .with_collection::<AudioAssets>()
-        .build(&mut app);
-    app.add_plugins(DefaultPlugins)
+    App::new()
+        .add_loopless_state(MyStates::AssetLoading)
+        .add_loading_state(
+            LoadingState::new(MyStates::AssetLoading)
+                .continue_to_state(MyStates::Next)
+                .with_collection::<TextureAssets>()
+                .with_collection::<AudioAssets>(),
+        )
+        .add_plugins(DefaultPlugins)
         // track progress during `MyStates::AssetLoading` and continue to `MyStates::Next` when progress is completed
         .add_plugin(ProgressPlugin::new(MyStates::AssetLoading))
         // gracefully quit the app when `MyStates::Next` is reached
