+++ bevy_asset_loader/src/asset_collection.rs
@@ -1,13 +1,15 @@
-use crate::asset_loader::DynamicAssets;
+use crate::dynamic_asset::DynamicAssets;
 use bevy::app::App;
 use bevy::asset::HandleUntyped;
 use bevy::prelude::World;
+
+pub use bevy_asset_loader_derive::AssetCollection;
 
 /// Trait to mark a struct as a collection of assets
 ///
 /// Derive is supported for structs with named fields.
 /// ```edition2021
-/// # use bevy_asset_loader::AssetCollection;
+/// # use bevy_asset_loader::prelude::*;
 /// # use bevy::prelude::*;
 /// #[derive(AssetCollection)]
 /// struct MyAssets {
@@ -66,12 +68,10 @@ pub trait AssetCollectionWorld {
 impl AssetCollectionWorld for World {
     fn init_collection<A: AssetCollection>(&mut self) {
         if self.get_resource::<A>().is_none() {
-            if self.get_resource::<DynamicAssets>().is_none() {
-                // This resource is required for loading a collection
-                // Since bevy_asset_loader does not have a "real" Plugin,
-                // we need to make sure the resource exists here
-                self.insert_resource(DynamicAssets::default());
-            }
+            // This resource is required for loading a collection
+            // Since bevy_asset_loader does not have a "real" Plugin,
+            // we need to make sure the resource exists here
+            self.init_resource::<DynamicAssets>();
             // make sure the assets start to load
             let _ = A::load(self);
             let collection = A::create(self);
