+++ bevy_asset_loader/src/lib.rs
@@ -5,7 +5,7 @@
 //! resources containing loaded handles and the plugin will switch to a second configurable [`State`](::bevy::ecs::schedule::State).
 //!
 //! ```edition2021
-//! # use bevy_asset_loader::{AssetLoader, AssetCollection};
+//! # use bevy_asset_loader::prelude::*;
 //! # use bevy::prelude::*;
 //! # use bevy::asset::AssetPlugin;
 //! #
@@ -14,20 +14,19 @@
 //!
 //! # #[cfg(not(feature="stageless"))]
 //! fn main() {
-//!     let mut app = App::new();
-//!     app
+//!     App::new()
 //! # /*
 //!         .add_plugins(DefaultPlugins)
 //! # */
 //! #       .add_plugins(MinimalPlugins)
 //! #       .init_resource::<iyes_progress::ProgressCounter>()
-//! #       .add_plugin(AssetPlugin::default());
-//!     AssetLoader::new(GameState::Loading)
-//!         .continue_to_state(GameState::Next)
-//!         .with_collection::<AudioAssets>()
-//!         .with_collection::<ImageAssets>()
-//!         .build(&mut app);
-//!     app
+//! #       .add_plugin(AssetPlugin::default())
+//!         .add_loading_state(
+//!             LoadingState::new(GameState::Loading)
+//!                 .continue_to_state(GameState::Next)
+//!                 .with_collection::<AudioAssets>()
+//!                 .with_collection::<ImageAssets>()
+//!         )
 //!         .add_state(GameState::Loading)
 //!         .add_system_set(SystemSet::on_update(GameState::Next)
 //!             .with_system(use_asset_handles)
@@ -38,24 +37,23 @@
 //!
 //! # #[cfg(all(feature="stageless"))]
 //! # fn main() {
-//! #     let mut app = App::new();
-//! #    app
+//! #     App::new()
 //! #       .add_loopless_state(GameState::Loading)
 //! # /*
 //!         .add_plugins(DefaultPlugins)
 //! # */
 //! #       .add_plugins(MinimalPlugins)
 //! #       .init_resource::<iyes_progress::ProgressCounter>()
-//! #       .add_plugin(AssetPlugin::default());
-//! #    AssetLoader::new(GameState::Loading)
-//! #        .continue_to_state(GameState::Next)
-//! #        .with_collection::<AudioAssets>()
-//! #        .with_collection::<ImageAssets>()
-//! #        .build(&mut app);
-//! #    app
-//! #        .add_system(use_asset_handles.run_in_state(GameState::Next))
-//! #        .set_runner(|mut app| app.schedule.run(&mut app.world))
-//! #        .run();
+//! #       .add_plugin(AssetPlugin::default())
+//! #       .add_loading_state(
+//! #         LoadingState::new(GameState::Loading)
+//! #           .continue_to_state(GameState::Next)
+//! #           .with_collection::<AudioAssets>()
+//! #           .with_collection::<ImageAssets>()
+//! #       )
+//! #       .add_system(use_asset_handles.run_in_state(GameState::Next))
+//! #       .set_runner(|mut app| app.schedule.run(&mut app.world))
+//! #       .run();
 //! }
 //!
 //! #[derive(AssetCollection)]
@@ -91,16 +89,27 @@
 #![warn(unused_imports, missing_docs)]
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
-mod asset_collection;
-mod asset_loader;
+/// Trait definition for types that represent a collection of assets
+///
+/// And extension traits to insert said collections into your Bevy app or world
+pub mod asset_collection;
+/// Types and infrastructure to load and use dynamic assets
+pub mod dynamic_asset;
+/// A game state responsible for loading assets
+pub mod loading_state;
 
-pub use crate::asset_collection::{AssetCollection, AssetCollectionApp, AssetCollectionWorld};
-pub use crate::asset_loader::{AssetLoader, DynamicAsset, DynamicAssets};
-
-#[cfg(feature = "dynamic_assets")]
-pub use crate::asset_loader::{DynamicAssetCollection, DynamicAssetCollections};
-
-pub use bevy_asset_loader_derive::AssetCollection;
+#[doc(hidden)]
+pub mod prelude {
+    #[doc(hidden)]
+    #[cfg(feature = "dynamic_assets")]
+    pub use crate::dynamic_asset::{DynamicAssetCollection, DynamicAssetCollections};
+    #[doc(hidden)]
+    pub use crate::{
+        asset_collection::{AssetCollection, AssetCollectionApp, AssetCollectionWorld},
+        dynamic_asset::{DynamicAsset, DynamicAssets},
+        loading_state::{LoadingState, LoadingStateAppExt},
+    };
+}
 
 #[cfg(all(feature = "2d", feature = "3d"))]
 #[doc = include_str!("../../README.md")]
