+++ bevy_asset_loader/src/loading_state/systems.rs
@@ -1,21 +1,20 @@
 use bevy::asset::{AssetServer, LoadState};
-use bevy::ecs::prelude::Commands;
-
+#[cfg(not(feature = "stageless"))]
+use bevy::ecs::prelude::State;
 use bevy::ecs::prelude::{FromWorld, World};
 use bevy::ecs::schedule::StateData;
 use bevy::ecs::system::SystemState;
-
+#[cfg(not(feature = "stageless"))]
+use bevy::ecs::world::WorldCell;
 use bevy::prelude::{Mut, Res, ResMut, Stage};
 use std::marker::PhantomData;
 
 #[cfg(feature = "progress_tracking")]
 use iyes_progress::{Progress, ProgressCounter};
 
-use iyes_loopless::prelude::{CurrentState, NextState};
-
 use crate::asset_collection::AssetCollection;
-use crate::asset_loader::{
-    AssetLoaderConfiguration, LoadingAssetHandles, LoadingState, LoadingStateSchedules,
+use crate::loading_state::{
+    AssetLoaderConfiguration, InternalLoadingState, LoadingAssetHandles, LoadingStateSchedules,
 };
 
 pub(crate) fn init_resource<Asset: FromWorld + Send + Sync + 'static>(world: &mut World) {
@@ -25,19 +24,17 @@ pub(crate) fn init_resource<Asset: FromWorld + Send + Sync + 'static>(world: &mu
 
 pub(crate) fn start_loading_collection<S: StateData, Assets: AssetCollection>(world: &mut World) {
     #[allow(clippy::type_complexity)]
-    let mut system_state: SystemState<(
-        ResMut<AssetLoaderConfiguration<S>>,
-        Res<CurrentState<S>>,
-    )> = SystemState::new(world);
+    let mut system_state: SystemState<(ResMut<AssetLoaderConfiguration<S>>, Res<State<S>>)> =
+        SystemState::new(world);
     let (mut asset_loader_configuration, state) = system_state.get_mut(world);
 
     let mut config = asset_loader_configuration
         .configuration
-        .get_mut(&state.0)
+        .get_mut(state.current())
         .unwrap_or_else(|| {
             panic!(
                 "Could not find a loading configuration for state {:?}",
-                state
+                state.current()
             )
         });
     config.loading_collections += 1;
@@ -49,15 +46,7 @@ pub(crate) fn start_loading_collection<S: StateData, Assets: AssetCollection>(wo
 }
 
 pub(crate) fn check_loading_collection<S: StateData, Assets: AssetCollection>(world: &mut World) {
-    if let Some((done, total, option_loading_collections)) =
-        count_loaded_handles::<S, Assets>(world)
-    {
-        if let Some(loading_collections) = option_loading_collections {
-            if loading_collections == 0 {
-                world.insert_resource(NextState(LoadingState::Finalize))
-            }
-        }
-
+    if let Some((done, total)) = count_loaded_handles::<S, Assets>(world.cell()) {
         if total == done {
             let asset_collection = Assets::create(world);
             world.insert_resource(asset_collection);
@@ -77,9 +66,8 @@ pub(crate) fn check_loading_collection<S: StateData, Assets: AssetCollection>(wo
 }
 
 fn count_loaded_handles<S: StateData, Assets: AssetCollection>(
-    world: &mut World,
-) -> Option<(u32, u32, Option<usize>)> {
-    let cell = world.cell();
+    cell: WorldCell,
+) -> Option<(u32, u32)> {
     let loading_asset_handles = cell.get_resource::<LoadingAssetHandles<Assets>>()?;
     let total = loading_asset_handles.handles.len();
 
@@ -94,45 +82,62 @@ fn count_loaded_handles<S: StateData, Assets: AssetCollection>(
         .filter(|state| state == &LoadState::Loaded)
         .count();
     if done < total {
-        return Some((done as u32, total as u32, None));
+        return Some((done as u32, total as u32));
     }
 
-    let mut loading_collections: Option<usize> = None;
-
     let state = cell
-        .get_resource::<CurrentState<S>>()
+        .get_resource::<State<S>>()
         .expect("Cannot get State resource");
+    let mut loading_state = cell
+        .get_resource_mut::<State<InternalLoadingState>>()
+        .expect("Cannot get LoadingStatePhase");
     let mut asset_loader_configuration = cell
         .get_resource_mut::<AssetLoaderConfiguration<S>>()
         .expect("Cannot get AssetLoaderConfiguration resource");
-    if let Some(mut config) = asset_loader_configuration.configuration.get_mut(&state.0) {
+    if let Some(mut config) = asset_loader_configuration
+        .configuration
+        .get_mut(state.current())
+    {
         config.loading_collections -= 1;
-        loading_collections = Some(config.loading_collections)
+        if config.loading_collections == 0 {
+            loading_state
+                .set(InternalLoadingState::Finalize)
+                .expect("Failed to set loading State");
+        }
     }
 
-    Some((done as u32, total as u32, loading_collections))
+    Some((done as u32, total as u32))
 }
 
-pub(crate) fn initialize_loading_state(mut commands: Commands) {
+pub(crate) fn initialize_loading_state(mut loading_state: ResMut<State<InternalLoadingState>>) {
     #[cfg(feature = "dynamic_assets")]
-    commands.insert_resource(NextState(LoadingState::LoadingDynamicAssetCollections));
+    loading_state
+        .set(InternalLoadingState::LoadingDynamicAssetCollections)
+        .expect("Failed to set LoadingState");
     #[cfg(not(feature = "dynamic_assets"))]
-    commands.insert_resource(NextState(LoadingState::LoadingAssets));
+    loading_state
+        .set(InternalLoadingState::LoadingAssets)
+        .expect("Failed to set LoadingState");
 }
 
 pub(crate) fn finish_loading_state<S: StateData>(
-    mut commands: Commands,
-    state: Res<CurrentState<S>>,
+    mut state: ResMut<State<S>>,
+    mut loading_state: ResMut<State<InternalLoadingState>>,
     asset_loader_configuration: Res<AssetLoaderConfiguration<S>>,
 ) {
-    if let Some(config) = asset_loader_configuration.configuration.get(&state.0) {
+    if let Some(config) = asset_loader_configuration
+        .configuration
+        .get(state.current())
+    {
         if let Some(next) = config.next.as_ref() {
-            commands.insert_resource(NextState(next.clone()));
+            state.set(next.clone()).expect("Failed to set next State");
             return;
         }
     }
 
-    commands.insert_resource(NextState(LoadingState::Done));
+    loading_state
+        .set(InternalLoadingState::Done)
+        .expect("Failed to set LoadingState");
 }
 
 pub(crate) fn run_loading_state<S: StateData>(world: &mut World) {
@@ -140,7 +145,7 @@ pub(crate) fn run_loading_state<S: StateData>(world: &mut World) {
         |world, mut loading_state_config: Mut<LoadingStateSchedules<S>>| {
             if let Some(schedule) = loading_state_config
                 .schedules
-                .get_mut(&world.get_resource::<CurrentState<S>>().unwrap().0)
+                .get_mut(world.get_resource::<State<S>>().unwrap().current())
             {
                 schedule.run(world);
             }
@@ -148,6 +153,7 @@ pub(crate) fn run_loading_state<S: StateData>(world: &mut World) {
     );
 }
 
-pub(crate) fn reset_loading_state(mut commands: Commands) {
-    commands.insert_resource(NextState(LoadingState::Initialize));
+pub(crate) fn reset_loading_state(mut state: ResMut<State<InternalLoadingState>>) {
+    // we can ignore the error, because it means we are already in the correct state
+    let _ = state.overwrite_set(InternalLoadingState::Initialize);
 }
