+++ bevy_asset_loader/tests/can_run_without_next_state.rs
@@ -4,7 +4,8 @@ use bevy::app::AppExit;
 use bevy::asset::AssetPlugin;
 use bevy::audio::AudioPlugin;
 use bevy::prelude::*;
-use bevy_asset_loader::{AssetCollection, AssetLoader};
+use bevy_asset_loader::asset_collection::AssetCollection;
+use bevy_asset_loader::loading_state::{LoadingState, LoadingStateAppExt};
 
 #[cfg_attr(
     all(
@@ -16,17 +17,13 @@ use bevy_asset_loader::{AssetCollection, AssetLoader};
     test
 )]
 fn can_run_without_next_state() {
-    let mut app = App::new();
-    app.add_state(MyStates::Load)
+    App::new()
+        .add_state(MyStates::Load)
         .add_plugins(MinimalPlugins)
         .add_plugin(AssetPlugin::default())
-        .add_plugin(AudioPlugin::default());
-
-    AssetLoader::new(MyStates::Load)
-        .with_collection::<MyAssets>()
-        .build(&mut app);
-
-    app.init_resource::<TestState>()
+        .add_plugin(AudioPlugin::default())
+        .add_loading_state(LoadingState::new(MyStates::Load).with_collection::<MyAssets>())
+        .init_resource::<TestState>()
         .add_system_set(
             SystemSet::on_update(MyStates::Load)
                 .with_system(timeout)
