+++ bevy_asset_loader/tests/stageless/init_resource.rs
@@ -4,23 +4,24 @@ use bevy::app::AppExit;
 use bevy::asset::AssetPlugin;
 use bevy::audio::AudioPlugin;
 use bevy::prelude::*;
-use bevy_asset_loader::{AssetCollection, AssetLoader};
+use bevy_asset_loader::asset_collection::AssetCollection;
+use bevy_asset_loader::loading_state::{LoadingState, LoadingStateAppExt};
 use iyes_loopless::prelude::*;
 
 #[test]
 fn init_resource() {
-    let mut app = App::new();
-    app.add_plugins(MinimalPlugins)
+    App::new()
+        .add_plugins(MinimalPlugins)
         .add_plugin(AssetPlugin::default())
         .add_plugin(AudioPlugin::default())
-        .add_loopless_state(MyStates::Load);
-    AssetLoader::new(MyStates::Load)
-        .continue_to_state(MyStates::Next)
-        .with_collection::<MyAssets>()
-        .init_resource::<PostProcessed>()
-        .build(&mut app);
-
-    app.add_system(timeout.run_in_state(MyStates::Load))
+        .add_loopless_state(MyStates::Load)
+        .add_loading_state(
+            LoadingState::new(MyStates::Load)
+                .continue_to_state(MyStates::Next)
+                .with_collection::<MyAssets>()
+                .init_resource::<PostProcessed>(),
+        )
+        .add_system(timeout.run_in_state(MyStates::Load))
         .add_enter_system(MyStates::Next, expect)
         .run();
 }
