+++ bevy_asset_loader/tests/ui/missing_feature.rs
@@ -1,4 +1,4 @@
-use bevy_asset_loader::*;
+use bevy_asset_loader::prelude::*;
 use bevy::prelude::*;
 
 fn main() {}
