+++ bevy_asset_loader/tests/ui_with_2d_3d_features/key_and_any_other_attribute.rs
@@ -1,4 +1,4 @@
-use bevy_asset_loader::*;
+use bevy_asset_loader::prelude::*;
 use bevy::prelude::*;
 
 fn main() {}
